Add files via upload
This class gave me the opportunity to do a semester's worth of research on examining various usability and privacy concerns that less tech-savvy individuals face on a day-to-day basis. Specifically, we were tasked with picking a potential technology could be widely used in 20-50 years, and examining those concerns. My group that I was a part of decided to do research on the use of ECAs (embodied conversational agents) in the medical field. This paper is our findings. A ton of credit goes to the graduate students that helped spearhead the writing of this paper (Alex Schlesener and Akshatha Arkit), as they were the individuals that had the most experience writing research papers compared to myself and the other undergraduate students who did not have similar experience.